refact: organize packages by domain
