+++ src/main/kotlin/com/dobby/backend/application/mapper/VerificationMapper.kt
@@ -1,7 +1,7 @@
 package com.dobby.backend.application.mapper
 
-import com.dobby.backend.application.usecase.signup.email.EmailCodeSendUseCase
-import com.dobby.backend.application.usecase.signup.email.EmailVerificationUseCase
+import com.dobby.backend.application.usecase.member.email.EmailCodeSendUseCase
+import com.dobby.backend.application.usecase.member.email.EmailVerificationUseCase
 import com.dobby.backend.infrastructure.database.entity.VerificationEntity
 import com.dobby.backend.infrastructure.database.entity.enum.VerificationStatus
 import com.dobby.backend.presentation.api.dto.request.signup.EmailSendRequest
