+++ src/main/kotlin/com/dobby/backend/application/service/SignupService.kt
@@ -1,8 +1,8 @@
 package com.dobby.backend.application.service
 
-import com.dobby.backend.application.usecase.signup.CreateParticipantUseCase
-import com.dobby.backend.application.usecase.signup.CreateResearcherUseCase
-import com.dobby.backend.application.usecase.signup.VerifyResearcherEmailUseCase
+import com.dobby.backend.application.usecase.member.CreateParticipantUseCase
+import com.dobby.backend.application.usecase.member.CreateResearcherUseCase
+import com.dobby.backend.application.usecase.member.VerifyResearcherEmailUseCase
 import com.dobby.backend.domain.exception.SignupOauthEmailDuplicateException
 import com.dobby.backend.domain.gateway.MemberGateway
 import com.dobby.backend.infrastructure.database.entity.enum.MemberStatus
