+++ src/main/kotlin/com/dobby/backend/presentation/api/controller/EmailController.kt
@@ -3,8 +3,8 @@ package com.dobby.backend.presentation.api.controller
 import com.dobby.backend.application.service.EmailService
 import com.dobby.backend.presentation.api.dto.request.signup.EmailSendRequest
 import com.dobby.backend.presentation.api.dto.request.signup.EmailVerificationRequest
-import com.dobby.backend.presentation.api.dto.response.signup.EmailSendResponse
-import com.dobby.backend.presentation.api.dto.response.signup.EmailVerificationResponse
+import com.dobby.backend.presentation.api.dto.response.member.EmailSendResponse
+import com.dobby.backend.presentation.api.dto.response.member.EmailVerificationResponse
 import com.dobby.backend.presentation.api.mapper.EmailMapper
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
