+++ src/main/kotlin/com/dobby/backend/presentation/api/dto/response/member/EmailVerificationResponse.kt
@@ -1,4 +1,4 @@
-package com.dobby.backend.presentation.api.dto.response.signup
+package com.dobby.backend.presentation.api.dto.response.member
 
 import io.swagger.v3.oas.annotations.media.Schema
