+++ src/main/kotlin/com/dobby/backend/presentation/api/mapper/EmailMapper.kt
@@ -1,11 +1,11 @@
 package com.dobby.backend.presentation.api.mapper
 
-import com.dobby.backend.application.usecase.signup.email.EmailCodeSendUseCase
-import com.dobby.backend.application.usecase.signup.email.EmailVerificationUseCase
+import com.dobby.backend.application.usecase.member.email.EmailCodeSendUseCase
+import com.dobby.backend.application.usecase.member.email.EmailVerificationUseCase
 import com.dobby.backend.presentation.api.dto.request.signup.EmailSendRequest
 import com.dobby.backend.presentation.api.dto.request.signup.EmailVerificationRequest
-import com.dobby.backend.presentation.api.dto.response.signup.EmailSendResponse
-import com.dobby.backend.presentation.api.dto.response.signup.EmailVerificationResponse
+import com.dobby.backend.presentation.api.dto.response.member.EmailSendResponse
+import com.dobby.backend.presentation.api.dto.response.member.EmailVerificationResponse
 
 object EmailMapper {
