+++ src/test/kotlin/com/dobby/backend/application/usecase/GenerateTestTokenUseCaseTest.kt
@@ -1,5 +1,6 @@
 package com.dobby.backend.application.usecase
 
+import com.dobby.backend.application.usecase.auth.GenerateTestTokenUseCase
 import com.dobby.backend.domain.gateway.MemberGateway
 import io.kotest.core.spec.style.BehaviorSpec
 import com.dobby.backend.domain.gateway.TokenGateway
