+++ src/test/kotlin/com/dobby/backend/application/usecase/GenerateTokenWithRefreshTokenUseCaseTest.kt
@@ -1,5 +1,6 @@
 package com.dobby.backend.application.usecase
 
+import com.dobby.backend.application.usecase.auth.GenerateTokenWithRefreshTokenUseCase
 import com.dobby.backend.domain.gateway.MemberGateway
 import com.dobby.backend.domain.gateway.TokenGateway
 import com.dobby.backend.domain.model.member.Member
